memcg, fsnotify: no oom-kill for remote memcg charging
d46eb14b735b ("fs: fsnotify: account fsnotify metadata to kmemcg") added
remote memcg charging for fanotify and inotify event objects.  The aim was
to charge the memory to the listener who is interested in the events but
without triggering the OOM killer.  Otherwise there would be security
concerns for the listener.  At the time, oom-kill trigger was not in the
charging path.  A parallel work added the oom-kill back to charging path
i.e.  commit 29ef680ae7c2 ("memcg, oom: move out_of_memory back to the
charge path").  So to not trigger oom-killer in the remote memcg,
explicitly add __GFP_RETRY_MAYFAIL to the fanotigy and inotify event
allocations.

Link: http://lkml.kernel.org/r/20190514212259.156585-2-shakeelb@google.com
Signed-off-by: Shakeel Butt <shakeelb@google.com>
Reviewed-by: Roman Gushchin <guro@fb.com>
Acked-by: Jan Kara <jack@suse.cz>
Cc: Johannes Weiner <hannes@cmpxchg.org>
Cc: Vladimir Davydov <vdavydov.dev@gmail.com>
Cc: Michal Hocko <mhocko@suse.com>
Cc: Amir Goldstein <amir73il@gmail.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>