resource: introduce resource cache
For efficient search of resources, as needed to determine the memory type
for dax page-faults, introduce a cache of the most recently used top-level
resource.  Caching the top-level should be safe as ranges in that level do
not overlap (unlike those of lower levels).

Keep the cache per-cpu to avoid possible contention.  Whenever a resource
is added, removed or changed, invalidate all the resources.  The
invalidation takes place when the resource_lock is taken for write,
preventing possible races.

This patch provides relatively small performance improvements over the
previous patch (~0.5% on sysbench), but can benefit systems with many
resources.

Link: http://lkml.kernel.org/r/20190613045903.4922-4-namit@vmware.com
Signed-off-by: Nadav Amit <namit@vmware.com>
Cc: Borislav Petkov <bp@suse.de>
Cc: Toshi Kani <toshi.kani@hpe.com>
Cc: Peter Zijlstra <peterz@infradead.org>
Cc: Dave Hansen <dave.hansen@linux.intel.com>
Cc: Dan Williams <dan.j.williams@intel.com>
Cc: Bjorn Helgaas <bhelgaas@google.com>
Cc: Ingo Molnar <mingo@kernel.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>