mm/z3fold.c: allow __GFP_HIGHMEM in z3fold_alloc
One of the gfp flags used to show that a page is movable is __GFP_HIGHMEM.
Currently z3fold_alloc() fails when __GFP_HIGHMEM is passed.  Now that
z3fold pages are movable, we allow __GFP_HIGHMEM.  We strip the movability
related flags from the call to kmem_cache_alloc() for our slots since it
is a kernel allocation.

[akpm@linux-foundation.org: coding-style fixes]
Link: http://lkml.kernel.org/r/20190712222118.108192-1-henryburns@google.com
Signed-off-by: Henry Burns <henryburns@google.com>
Acked-by: Vitaly Wool <vitalywool@gmail.com>
Reviewed-by: Shakeel Butt <shakeelb@google.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>