mm,sparse: fix deactivate_section for early sections
deactivate_section checks whether a section is early or not in order to
either call free_map_bootmem() or depopulate_section_memmap().  Being the
former for sections added at boot time, and the latter for sections
hotplugged.

The problem is that we zero section_mem_map, so the last early_section()
will always report false and the section will not be removed.

Fix this checking whether a section is early or not at function
entry.

Link: http://lkml.kernel.org/r/20190715081549.32577-2-osalvador@suse.de
Fixes: mmotm ("mm/sparsemem: Support sub-section hotplug")
Signed-off-by: Oscar Salvador <osalvador@suse.de>
Reviewed-by: Aneesh Kumar K.V <aneesh.kumar@linux.ibm.com>
Cc: Dan Williams <dan.j.williams@intel.com>
Cc: Michal Hocko <mhocko@suse.com>
Cc: Vlastimil Babka <vbabka@suse.cz>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>