mm,memory_hotplug: fix shrink_{zone,node}_span
Since [1], shrink_{zone,node}_span work on PAGES_PER_SUBSECTION
granularity.  We need to adapt the loop that checks whether a zone/node
contains only holes, and skip the whole range to be removed.

Otherwise, since sub-sections belonging to the range to be removed have
not yet been deactivated, pfn_valid() will return true on those and we
will be left with a wrong accounting of spanned_pages, both for the zone
and the node.

Link: http://lkml.kernel.org/r/20190717090725.23618-3-osalvador@suse.de
Fixes: mmotm ("mm/hotplug: prepare shrink_{zone, pgdat}_span for sub-section removal")
Signed-off-by: Oscar Salvador <osalvador@suse.de>
Reviewed-by: Dan Williams <dan.j.williams@intel.com>
Cc: Aneesh Kumar K.V <aneesh.kumar@linux.ibm.com>
Cc: Michal Hocko <mhocko@suse.com>
Cc: Vlastimil Babka <vbabka@suse.cz>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>