mm/gup.c: introduce __put_user_pages()
Patch series "introduce __put_user_pages(), convert a few call sites"

As discussed in [1], this adds a more capable variation of
put_user_pages() to the API set, and uses it to simplify both the main
implementation, and (especially) the call sites.

Thanks to Christoph for the simplifying ideas, and Matthew for (again)
recommending an enum in the API.  Matthew, I seem to recall you asked for
enums before this, so I'm sorry it took until now for me to add them.  :)

The new __put_user_pages() takes an enum that handles the various
combinations of needing to call set_page_dirty() or set_page_dirty_lock(),
before calling put_user_page().

Notes about the remaining patches to come:

There are about 50+ patches in my tree [2], and I'll be sending out the
remaining ones in a few more groups:

    * The block/bio related changes (Jerome mostly wrote those, but I've
      had to move stuff around extensively, and add a little code)

    * mm/ changes

    * other subsystem patches

    * an RFC that shows the current state of the tracking patch set. That
      can only be applied after all call sites are converted, but it's
      good to get an early look at it.

This is part a tree-wide conversion, as described in fc1d8e7cca2d ("mm:
introduce put_user_page*(), placeholder versions").

[1] https://lore.kernel.org/r/20190722093355.GB29538@lst.de
[2] https://github.com/johnhubbard/linux/tree/gup_dma_core


This patch (of 3):

Add a more capable variation of put_user_pages() to the API set, and call
it from the simple ones.

The new __put_user_pages() takes an enum that handles the various
combinations of needing to call set_page_dirty() or set_page_dirty_lock(),
before calling put_user_page().

Link: http://lkml.kernel.org/r/20190722223415.13269-2-jhubbard@nvidia.com
Signed-off-by: John Hubbard <jhubbard@nvidia.com>
Reviewed-by: Andrew Morton <akpm@linux-foundation.org>
Cc: Ira Weiny <ira.weiny@intel.com>
Cc: Matthew Wilcox <willy@infradead.org>
Cc: Jan Kara <jack@suse.cz>
Cc: Christoph Hellwig <hch@lst.de>
Cc: Alexander Viro <viro@zeniv.linux.org.uk>
Cc: Boaz Harrosh <boaz@plexistor.com>
Cc: Daniel Vetter <daniel@ffwll.ch>
Cc: Dan Williams <dan.j.williams@intel.com>
Cc: Dave Chinner <david@fromorbit.com>
Cc: David Airlie <airlied@linux.ie>
Cc: "David S . Miller" <davem@davemloft.net>
Cc: Ilya Dryomov <idryomov@gmail.com>
Cc: Jason Gunthorpe <jgg@ziepe.ca>
Cc: Jens Axboe <axboe@kernel.dk>
Cc: Jerome Glisse <jglisse@redhat.com>
Cc: Johannes Thumshirn <jthumshirn@suse.de>
Cc: Magnus Karlsson <magnus.karlsson@intel.com>
Cc: Miklos Szeredi <miklos@szeredi.hu>
Cc: Ming Lei <ming.lei@redhat.com>
Cc: Sage Weil <sage@redhat.com>
Cc: Santosh Shilimkar <santosh.shilimkar@oracle.com>
Cc: Yan Zheng <zyan@redhat.com>
Cc: Björn Töpel <bjorn.topel@intel.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>