string: add stracpy() and stracpy_pad() mechanisms
Several uses of strlcpy and strscpy have had defects because the last
argument of each function is misused or typoed.

Add macro mechanisms to avoid this defect.

stracpy (copy a string to a string array) must have a string array as the
first argument (to) and uses sizeof(to) as the size.

These mechanisms verify that the to argument is an array of char or other
compatible types like u8 or unsigned char.

A BUILD_BUG is emitted when the type of to is not compatible.

Link: http://lkml.kernel.org/r/7ab8957eaf9b0931a59eff6e2bd8c5169f2f6c41.1563841972.git.joe@perches.com
Signed-off-by: Joe Perches <joe@perches.com>
Cc: Jann Horn <jannh@google.com>
Cc: Jonathan Corbet <corbet@lwn.net>
Cc: Kees Cook <keescook@chromium.org>
Cc: Nitin Gote <nitin.r.gote@intel.com>
Cc: Rasmus Villemoes <rasmus.villemoes@prevas.dk>
Cc: Stephen Kitt <steve@sk2.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>