mips: mm: add p?d_leaf() definitions
walk_page_range() is going to be allowed to walk page tables other than
those of user space.  For this it needs to know when it has reached a
'leaf' entry in the page tables.  This information is provided by the
p?d_leaf() functions/macros.

For mips, we only support large pages on 64 bit.

For 64 bit if _PAGE_HUGE is defined we can simply look for it.  When not
defined we can be confident that there are no leaf pages in existence and
fall back on the generic implementation (added in a later patch) which
returns 0.

Link: http://lkml.kernel.org/r/20190722154210.42799-5-steven.price@arm.com
Signed-off-by: Steven Price <steven.price@arm.com>
Cc: Ralf Baechle <ralf@linux-mips.org>
Cc: Paul Burton <paul.burton@mips.com>
Cc: James Hogan <jhogan@kernel.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>