x86: mm: convert ptdump_walk_pgd_level_core() to take an mm_struct
An mm_struct is needed to enable x86 to use of the generic
walk_page_range() function.

In the case of walking the user page tables (when
CONFIG_PAGE_TABLE_ISOLATION is enabled), it is necessary to create a
fake_mm structure because there isn't an mm_struct with a pointer to the
pgd of the user page tables.  This fake_mm structure is initialised with
the minimum necessary for the generic page walk code.

Link: http://lkml.kernel.org/r/20190722154210.42799-19-steven.price@arm.com
Signed-off-by: Steven Price <steven.price@arm.com>
Cc: Ingo Molnar <mingo@elte.hu>
Cc: "H. Peter Anvin" <hpa@zytor.com>
Cc: Thomas Gleixner <tglx@linutronix.de>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>